fix: Vega armv7 cross-compilation errors
Make diagnostic attribute construction explicit so 32-bit integer types
like armv7 size_t do not hit ambiguous std::variant construction.

Also make timestamp storage and upload backoff min comparisons use the
int64_t type expected by the SDK.

+++ src/datadog/cpp/attribute.cpp
@@ -145,7 +145,9 @@ Attribute Attribute::Int(int64_t value) { return Attribute(ValueType::Int, value
 Attribute Attribute::UInt(uint64_t value) { return Attribute(ValueType::UInt, value); }
 
 Attribute Attribute::Timestamp(datadog::Timestamp value) {
-  return Attribute(ValueType::Timestamp, value.time_since_epoch().count());
+  return Attribute(
+      ValueType::Timestamp, static_cast<int64_t>(value.time_since_epoch().count())
+  );
 }
 
 Attribute Attribute::Double(double value) {
